refactor: cleanup API routes into different files (#25)
* feat(package): add migration script
