+++ packages/server/src/routes/api/categories/getCategories.ts
@@ -1,0 +1,18 @@
+import type { Request, Response } from "express";
+import { prisma } from "../../../lib/db";
+
+export const getCategories = async (req: Request, res: Response) => {
+  try {
+    const categories = await prisma.category.findMany({
+      include: {
+        _count: {
+          select: { items: true },
+        },
+      },
+      orderBy: { name: "asc" },
+    });
+    res.json(categories);
+  } catch (error) {
+    res.status(500).json({ error: "Failed to fetch categories" });
+  }
+};
