+++ packages/server/src/routes/api/index.ts
@@ -1,0 +1,43 @@
+// Stats
+export { getStats } from "./stats/getStats";
+
+// Scrapers
+export { getScrapers } from "./scrapers/getScrapers";
+export { scrapeScraper } from "./scrapers/scrapeScraper";
+export { enableScraper } from "./scrapers/enableScraper";
+export { disableScraper } from "./scrapers/disableScraper";
+
+// Websites
+export { getWebsites } from "./websites/getWebsites";
+export { addWebsite } from "./websites/addWebsite";
+export { deleteWebsite } from "./websites/deleteWebsite";
+export { triggerScrapeWebsite } from "./websites/triggerScrapeWebsite";
+export { scrapeBopa } from "./websites/scrapeBopa";
+
+// Auctions
+export { getAuctions } from "./auctions/getAuctions";
+export { getAuction } from "./auctions/getAuction";
+export { triggerScrape } from "./auctions/triggerScrape";
+
+// Items
+export { getItems } from "./items/getItems";
+export { getItemCategoryProbabilities } from "./items/getItemCategoryProbabilities";
+
+// Categories
+export { getCategories } from "./categories/getCategories";
+export { getCategory } from "./categories/getCategory";
+export { createCategory } from "./categories/createCategory";
+export { updateCategory } from "./categories/updateCategory";
+export { deleteCategory } from "./categories/deleteCategory";
+export { getItemsByCategory } from "./categories/getItemsByCategory";
+export { setItemMainCategory } from "./categories/setItemMainCategory";
+export { saveItemCategoryProbabilities } from "./categories/saveItemCategoryProbabilities";
+
+// AI Categorization
+export { categorizeItem } from "./categorization/categorizeItem";
+export { categorizeItems } from "./categorization/categorizeItems";
+export { categorizeAuction } from "./categorization/categorizeAuction";
+export { getItemCategorization } from "./categorization/getItemCategorization";
+
+// Database Management
+export { wipeTables } from "./database/wipeTables";
